% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{community_mobility_changes}
\alias{community_mobility_changes}
\title{Community Mobility Changes}
\format{
A data frame with 3 variables.
\describe{
\item{country}{Country name}
\item{community_mobility}{5 Place categories:Grocery & pharmacy, Parks, Residential, Retail & Recreation,
Transit stations, Workplaces}
\item{mean}{Mean value of mobility changes compared with date before the pandemic
reported as a positive or negative percentage.}
}
}
\source{
Google COVID-19 Community Mobility Reports \href{https://www.google.com/covid19/mobility/}{website}.
}
\usage{
community_mobility_changes
}
\description{
Data chart movement trends over time
by geography, across different categories of places such as retail and
recreation, groceries and pharmacies, parks, transit stations,
workplaces, and residential" to help
interested parties to assess responses to social distancing guidance
related to Covid-19.
}
\keyword{datasets}
