% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid19_cumulative_case}
\alias{covid19_cumulative_case}
\title{COVID-19 Cumulative Cases}
\format{
A data frame with 7 variables.
\describe{
\item{country}{Country name}
\item{iso3c}{ISO3c country code as defined by ISO 3166-1 alpha-3}
\item{date}{Calendar date in YYYY-MM-DD format}
\item{confirmed}{Confirmed Covid-19 cases as reported by JHU CSSE (accumulated)}
\item{deaths}{Covid-19-related deaths as reported by JHU CSSE (accumulated)}
\item{recovered}{Covid-19 recoveries as reported by JHU CSSE (accumulated)}
\item{cum_cases}{Cumulative confirmed Covid-19 cases grouped by countries}
}
}
\source{
Johns Hopkins University Center for Systems Science and
Engineering (JHU CCSE) Coronavirus Repository \href{https://github.com/CSSEGISandData/COVID-19}{website}.
}
\usage{
covid19_cumulative_case
}
\description{
COVID-19 confirmed cases, deaths, recovered cases and cumulative confirmed cases based on JHU CSSE data.
}
\keyword{datasets}
