% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid19_time}
\alias{covid19_time}
\title{COVID-19 Cases in Long Format}
\format{
A data frame with 6 variables.
\describe{
\item{country}{Country name}
\item{date}{Calendar date in YYYY-MM-DD format}
\item{month}{Month of date}
\item{day}{Day of date}
\item{type}{Type of reported cases including confirmed cases, reported deaths and recovered cases}
\item{number}{Number of cases for each type}
}
}
\source{
Johns Hopkins University Center for Systems Science and
Engineering (JHU CCSE) Coronavirus Repository \href{https://github.com/CSSEGISandData/COVID-19}{website}.
}
\usage{
covid19_time
}
\description{
Long format of COVID-19 cases with month and day.
}
\keyword{datasets}
