% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtrack-package.R
\docType{package}
\name{covidtrack-package}
\alias{covidtrack}
\alias{covidtrack-package}
\title{covidtrack: Track COVID-19 cases}
\description{
To create a Shiny App to show the COVID-19 cases and community mobility changes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etc5523-2020/r-package-assessment-yezihe-0063}
  \item Report bugs at \url{https://github.com/etc5523-2020/r-package-assessment-yezihe-0063/issues}
}

}
\author{
\strong{Maintainer}: Yezi He \email{yhee0063@student.monash.edu}

}
\keyword{internal}
