% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter data by countries and cases types}
\usage{
filter_data(data, country_name, type_name)
}
\arguments{
\item{data}{A data frame containing a \code{country} and \code{type} variable.}

\item{country_name}{A country you want to select.}

\item{type_name}{Types you can select including "confirmed", "deaths" and "recovered".
If you want to multiple select types, please use c("type1", "type2").}
}
\description{
A function to filter the data by countries and cases types.
}
\examples{
\dontrun{
covid19_time_Australia <- filter_data(covid19_time, "Australia",c("confirmed","deaths"))
tibble::view(covid19_time_Australia)
}

}
