% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jhu_covid19_data}
\alias{jhu_covid19_data}
\title{Johns Hopkins University CSSE data on COVID-19}
\format{
A data frame with 7 variables.
\describe{
\item{country}{Country name}
\item{iso3c}{ISO3c country code as defined by ISO 3166-1 alpha-3}
\item{date}{Calendar date in YYYY-MM-DD format}
\item{confirmed}{Confirmed Covid-19 cases as reported by JHU CSSE (accumulated)}
\item{deaths}{Covid-19-related deaths as reported by JHU CSSE (accumulated)}
\item{recovered}{Covid-19 recoveries as reported by JHU CSSE (accumulated)}
\item{timestamp}{Date and time where data has been collected from authoritative sources}
}
}
\source{
Johns Hopkins University Center for Systems Science and
Engineering (JHU CCSE) Coronavirus Repository \href{https://github.com/CSSEGISandData/COVID-19}{website}.
}
\usage{
jhu_covid19_data
}
\description{
Johns Hopkins University CSSE data on the spread of the SARS-CoV-2 virus and the COVID-19 pandemic.
}
\keyword{datasets}
