% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_country.R
\name{select_country}
\alias{select_country}
\title{Select a country for shiny app}
\usage{
select_country(data)
}
\arguments{
\item{data}{A data frame containing a \code{country} variable.}
}
\description{
A slider to select a country which you want to display.
}
\examples{
select_country(jhu_covid19_data)

}
